requirements.txt
streamlit
pdfplumber
python-dateutil
docx2txt
regex